requirements.txt
aiocache==0.12.3
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.109.2
fastapi-cli==0.0.7
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.16
pycparser==2.22
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
Pygments==2.19.1
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
qqmusic-api-python==0.3.4
rich==14.0.0
rich-toolkit==0.14.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.36.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.0
uvicorn==0.27.1
watchfiles==1.0.4
websockets==15.0.1
